# Assignments/FinalProject/swagger.yaml
openapi: 3.0.1
info:
  title: Bookstore Management API
  description: RESTful API for managing authors, books, customers, and orders with sales reporting (it is not very reliable please use POSTMAN and the testing file provided within this project).
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /orders:
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        201:
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      summary: Retrieve all orders
      responses:
        200:
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
...
  /reports:
    get:
      summary: Retrieve sales reports by date range
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        200:
          description: List of sales reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesReport'
components:
  schemas:
    Author:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        bio:
          type: string
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          $ref: '#/components/schemas/Author'
        genres:
          type: array
          items:
            type: string
        published_at:
          type: string
          format: date-time
        price:
          type: number
        stock:
          type: integer
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            postal_code:
              type: string
            country:
              type: string
    Order:
      type: object
      properties:
        id:
          type: integer
        customer:
          $ref: '#/components/schemas/Customer'
        items:
          type: array
          items:
            type: object
            properties:
              book:
                $ref: '#/components/schemas/Book'
              quantity:
                type: integer
        total_price:
          type: number
        status:
          type: string
        created_at:
          type: string
          format: date-time
    SalesReport:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        total_revenue:
          type: number
        total_orders:
          type: integer
        top_selling_books:
          type: array
          items:
            type: object
            properties:
              book:
                $ref: '#/components/schemas/Book'
              quantity_sold:
                type: integer